Solution Day 5 2021
Really should have written more tests for this one to drive out the `points` methods as I solved this one with trial and error and debug.  Lot's of fiddling the values used in the ranges for the x coordinate, the choice to increment vs. decrement the y coordinate, and which point from the vent definition to use as the y coordinate base.

Also I had to look up some geometry :(.  TBH fairly surprised that the radians -> degrees code actually produces fully whole numbers.