Solution for Day 7 2021
Felt weirdly simpler than day 6, so I wonder if I missed something about the others, or I just had a spark of intuition about it that meant I could crack it out quickly.  The longest time I spent was looking for the formula for the sum of natural numbers up to n `(n * (n+1))/2` because I couldn't remember the name of it and got lost in a wiki-hole about mathematical summations of infinite series.