Solution Day 8 2021
First part was pretty easy, but took me a while to understand the actual problem domain which wasn't actually required for this part of the problem.  Had to sit and work out a way to determine the segments for part 2 and that took a while, fun, but took a while.  Keen to see other approaches as I don't know my logic is the most efficient.