Solution Day 11 2021
Can't help but feel that my approach of iterating several times to calculate the flashes and energy changes is inefficient.  It feels like there's _probably_ some quick maths to be done using the rules of the system to determine in one pass how much extra energy a given cell will get and if it'll flash.  Still, this works, and even for ruby it's not tooo slow.