Solution Day 13 2021
Turns out, my solution for part 1 only passes by mistake, the horizontal fold algorithm I'm using is folding the wrong way.  One of the folds in the input isn't down the middle, whereas all the folds in the examples are.  This means for one of the folds if we count backwards we put things in the wrong place.  For example if we split [0,1,2,x,4,5] on the x, then 2 goes with 4 and 1 goes with 5 and 0 goes with nothing, but by counting backwards I had 0 gets 5, 1 gets 4 and 2 gets nothing.  I spent a lot of time thinking the problem was how I was constructing the paper in the first place rather than thinking the algorithm was wrong, because I assumed all the folds would be along halfs.  In the end I zipped the base and fold arrays together with a reverse and this ties them up right.

I've left the error in solution-1 to "show my working".  I suspect there's a tighter solution that wouldn't have fallen into this trap in the first place.