Solution Day 14 2021
Another one where part 2 is a "you're going to have to re-think this for perf reasons" iteration on part 1.  What's interesting about this one (at least my implementation of it) is that, unlike Day 6 with the fish, I don't think I can construct the world anymore.  With the fish I represent the problem as a count of how many are at each energy level and this means I can represent all the information I have - the total number of fish.  With the polymer though, I can't actually construct the polymer from the counts.  I'm only storing the count of each pair and the count of each element, there's no way to put use those counts to reconstruct the actual polymer in the correct order of pairs as they would be in from the generation process.

I dunno.  Interesting.