Solution Day 16 2021
On reading I wondered if I'd need to use string#pack, but I didn't.  I do wonder if all my casting from string to int via bases is actually slow and maybe there's a version where I do direct bit manipulation.  But... it's not actually a slow process to run on the provided input so I'm not that bothered.  Actually enjoyed this one quite a bit.