Solution Day 17 2021
Feel like I was probably supposed to do some kind of maths to work out the curves, but I just brute forced it.  In particular my limit for what to look at when calculating the y velocities to try was based on getting the answer wrong, and just doing some exploratory testing to see what values I was missing out by halting early.  I then used the same idea for calculating the x velocities on part 2 because for part 1 I only cared about x velocities that would 0 out within the target, but for part 2 we want ones that would 0 out afterwards.