Solution Day 18 2021
Start this in the morning and thought I'd represent things with actual nested pairs, but while this made it simple to do magnitude which is explicitly called out as recursive in the description, it turned out to be really hard to do the explode.  Came back to it in the evening and tried out a different representation with an array of numbers and brackets and using a stack style approach.  It's not particularly elegant, with all the break-based loops, but it works.  Part 2 was fairly quick after getting Part 1 done finally.