Solution Day 20 2021
Thought I had this working on the examples, but when I tried it on the input it just refused to get the right answer no matter what I tweaked.  Struggled to make headway on my own so I looked on reddit to get some pointers.  Turns out the problem isn't how much you expand / contract the result by, but the understanding of what the infinite-ness of the world means.  If, as in the input, the 0th element of the algorithm is '#' and the 511th element is '.' then the infinite plane flips every step from '#' to '.' so you need to encode that in however you expand the image and/or enhance the pixels based on their neighbours.  It does mean that the answer for every 2nd step of "how many lights are lit" is actually "infinite" :/

ps. Oh, but where's day 19.  Yeah, I just couldn't get my head round that one (pun intended).  3D maths made my head hurt and I just couldn't get started with it, so I left it.  Perhaps now I've broken the seal on getting hints from reddit I may return to it.  We'll see what Day 21 brings.