Solution Day 21 2021
Another "you're gonna need to model this completely differently for performance reasons in part 2" type of puzzle, but also it felt like a massive change in complexity vs. Day 20 or 19.  Which, I think, is revealed by the global stats.

My solution, as per, was to model the game completely with players for Part 1, and then in Part 2 to use a hash of counts of unique games.  This way we only need to iterate over a few thousand games per "tick".  I suspect that there's some kind of calculation that can be done to model the game in a single tick because the position + score changes are knowable with the deterministic dice, and with the quantum dirac dice.

Not smart enough to know this though, and my solutions do actually run in reasonable time.  Enjoyed this _waaay_ more than Day 20 (almost certainly because I solved it solo and there were no gotchas in the puzzle).