2023 Day 2 part 2: Minor tweak
I didn't like the `[nil, 0]` part, or the repeated `colour = cube_string.match ...` so I remembered about using `scan` for repeating patterns and regexp golfed my way out.

Arguably, the previous version was clearer, but this is _probably_ a bit faster.

I had thought to replace the final `{blue: blue, ...}` line with value-less hash syntax `{blue:, ...}` but in the end, I got rid of that line entirely.

Still not sure this is _better_ but this kind of "one-liner-ism" feels a bit advent-of-code-y so ... 🤷