2023 Day 3 part 2: Model "numbers" and "symbols" with Data.define
This lets us drastically reduce the number of "ranges" we create for each iteration through the space to once per "number" instance, rather than doing it several times in each calculation.  This makes both the original "numbers * potential_gears" gears calculation, and the "potential_gear_parts & potential_gears" calculations faster.  However, the original goes back to being fastest, proving that the improvement from the last one, was just about how many of those range objects we create.  Now we remove that from the picture, it is just about how many times we compare a number to a symbol, and the first implementation wins on that.

There's probably something to be said for improving the algorithm elsewhere to not iterate so many times, or to somehow consider significantly fewer symbols per number or vice versa.  But it's pretty fast now (~2.something seconds down to ~1ish second).

Interesting things I learned about `Data.define` - if you want to create methods and memoize them, you have to assign to instance variables in initialize as you can't modify the instance at all - even to assign to new instance variables.  You also have to do that _before_ you call out to `super` as that's what freezes the instance.  Makes sense once you think about it, but surprised me a bunch.