2023 Day 4 part 2 - first pass, too slow
At least in ruby this simplistic approach is too slow (e.g. on the input I haven't seen it complete).  I suspect the problem is that `to_see` is scaling hugely and I'm just eating too much memory with lots of repeated effort.  Let's see what we can do to speed this up.