2023 Day 4 part 2 - optimisation
As ever, use some data structures and exploit the game rules.  We know we'll never win a card with a lower id from a card so we only need to iterate through once.  Once we've done card 4 we never revisit and never backtrack.  So let's use a hash of id -> count.  We go through each card and find out the prize cards, then add current card count to the count for each prize.  This way, a single iteration over the cards and we're done.