2023 Day 5 part 1
Had to reimplement for perf reasons for example vs. input in part 1 which doesn't feel great.  Originally I had a hash of source -> destination based on the maps with a default_proc of returning the source number.  Obvs, with the scale of the numbers in the input that meant _huge_ arrays, so I shifted to a proc based approach - where we build ranges to cover the source and procs to calculate the destination value.

Walking the maps is a bit hairy, and I'm sure part 2 is going to require more effort around that.