2023 Day 2 part 2
WIP - I had a solution that worked, but looked like it would take ~20hrs to run so I attempted to optimise.  I _thought_ I could instead of having a map of x->y funcs that I walk through for each part of the path from seed to location, I could instead calculate the change for the whole map for seed -> location and have a single map.  This map would be made of ranges and I'd split each range based on the ranges in the next part of the map.  Turns out, that took me forever to work out and in the end was wrong.

I had a map of what changes each layer would make to the same number, however, if seed is 10 and that maps to soil 12 then I'd have 2 as the change, then I'd get the fertilizer change for 10 which might be 9 and I'd have 1 as the change.  seed -> fertilizer for 10 becomes a total change of +2 - 1 = 1 e.g. seed -> fertilized becomes 11.  That's _wrong_ though because at the fertilizer I don't need the change for 10, I need the change for 12.  All that mental effort and it didn't even work.  It was faster, looking like it'd take ~30mins to solve, but the result wouldn't be correct so what's the point.

Ultimately, I had focussed on the wrong part of the problem.  It's not that I have to do 7 func calls per seed, it's that I have to look at 1,975,502,102 seeds.  Any real perf will come from looking at less seeds somehow, not from doing less per seed.