2023 Day 5 part 2
Rather than trying to solve the problem by reducing the number of things to do, instead I chose to just brute force it with the parallel gem - I've got 8 cores so the ~20hr version should run in ~2.5hrs.  In the end, I couldn't work out how to get my `Map#each_seed` lazy yielder to run in parallel without creating a full array which I didn't want to do.  So in the end I just used the array I did have `seeds` which means asking each process not to spin on single seeds, but to spin on single ranges.  Means the progressbar updates every 30mins or so as a whole range is completed, and there are period where less than 8 processes are running to complete the single range they have.  But eh, it worked.

In talking about it at work I did come to the realisation that a possible solution might be:

Take the ranges for seed then smash them into the ranges for each part of the map in turn and split them.  Eventually we'll get all the slices of numbers that will require different paths.  We can then get location for the first / last of each of those and pick the smallest - every number in the range _should_ get the same transformation.

Can't be bothered coding it up though now.  I've got the answer and my 2nd star.