2023 Day 6
Uh.  What?

Day 5 - a hellish nightmare.  Day 6 - simple math problem.  Perhaps now I'm thinking AoC-style because although the temptation was still to model everything the task describes, it felt clear to me that the input for part 1 or 2 would give me big numbers and I couldn't just work out all the strategies and then pick the big ones.

I worked out, perhaps thinking along the same lines as the previous day's solution that I didn't bother implementing, that I could find the smallest and largest numbers and then return the diff between them.  If holding for 4ms is the smallest record breaker, and holding for 10ms is the biggest record breaker, then any number inbetween will also win - we can exploit the continuous nature of things.

So I did write the "walk all the numbers and generate the strategies" just in case, but of course, I didn't need them.

Pleasantly surprised that all I had to mess with for part 2 vs part 1 was the "Normalizer" routine.