2023 Day 7 - tidy up jokered type calc
Not by much, but in most cases we don't need to check the count of J's as the change is always the same.  Now we've run the solution on the input txt we are pretty sure we've not missed any examples either, so we don't need the else raise bits.