2023 Day 8
Part 1 was fairly straightforward, as ever I spent a lot of time modelling features of the Map that I didn't need, but they helped me understand what was going on and debug as I invariably made trivial errors in the implementation.

Part 2 I started (see MultiMap) with a similar version to Part 1 that just did what the text says - walk the nodes from each starting position simultaneously until all paths have reached a terminal.  _Oooooohbveeeusssleee_ that was going to take forever.  So I scratched my head.

I ran my part 1 solution tweaked a bit to specify start + end and noticed that some starts and ends combos returned a path quickly, others didn't.  But so far, of the ones I tried, none were repeats.  That got me thinking, but not enough so I looked for some hints on reddit and that gave me the hint I needed - the paths are all loops, so you just need to work out the path for each **A to its first + only **Z then you find the lowest common multiple of all the path lengths.

This isn't a general solution, because there's no guarantee all the paths in the input would be loops.  However, these paths _are_ loops so it works.  They're also not _massive_ loops so it runs in pretty fast time.

The trick here is that you have to examine not just the puzzle description, but also the puzzle input to look for the tricks that'll get you the solution to this input, not the solution to the general problem.

Neat thing I learned (/ remembered) is that ruby's `reduce` doesn't have "special" understanding of :+ or :*, it's treating it as a almost but not quite Symbol#to_proc and calling `elem.send(<symbol>, next_elem)` (ish).  So you can use any method on your objects that take a single arg.  Like `Integer#lcm`!