2023 Day 10
I thought this would be more painful, but I remembered that checking if something is inside a loop is counting the crossings on the line, if the crossings before the point you care about is odd then it's inside the loop, if it's even, it's outside the loop.  The trick is ignoring the `-` pipes and working out if `F` is followed by `7` which makes it not a crossing, or `J` which means it is a crossing (and similar for `L` + `J` vs `L` + `7`).  That made part 2 quite simple.  I think part 1 I've probably not optimised, and have implemented things I don't need (like render_pipe_loop) - but they did make part 2 simpler as I could ignore any "extra" pipe parts.