2023 Day 11
I guess I'm just running a day behind now.  This wasn't too hard either.  Again, I think I spotted quickly that although it's compelling to create the expanded universe and render it, you don't need to at all.  The fact that _distance_ is just a simple x and y calculation should have pointed me to the part 2 solution in part 1.  I don't think the code is all that messy, but it doesn't fill me with joy that I got the answers for some of my part 2 specs by assuming the code worked and finding out what the test failures were and copying the "got" value into the "expected" value.