2023 Day 12 part 1
This ran pretty slowly, but worked.  Very clear though that part 2 requires _an algorithm_.  For the last few parts I've spotted the algorithm early on and so not hit the perf problems that part 2 scales you too.  I haven't spotted the algorithm for this one so my "just recurse on everything" plan isn't going to scale.  I'll let it stew in my brain and see what I can come up with later.