2023 Day 12 Part 2
NGL - I looked for help on this one.  I realised there was a lot of repeated work in my part 1 solution and I had to do 2 things:

1. reduce the amount of cases I considered - we don't need to generate all the possibilities only to then throw some away later - we can reject possibilities as we generate them
2. the solution ends up iterating over possibility ranges it has looked at before and so we can memoize those results

I started by trying to just add a cache to generating the possibilities - that worked, but didn't help with perf for part 2.

Then I started trying to look at rejecting possibilities but that just caused me data structure woes.  I tried collecting the part of the string we've looked at and stopping if the groups from that didn't match the groups we were trying.

In the end, I went to reddit and followed someone's pretty good walkthrough in python.  This approach looks per character, and recuses, like my attempts, but is also popping the stack of groups and will advance the cursor to claim a whole group or reject any future recursion if the current character is a #.  That's the massive step change in perf as it means you leap through the map at group scale and slice off whole swathes of the tree in one go.

Can't say I'm too happy with my brain for not getting there, but it made me think a bit and it was nice to revisit some python.  Loved learning that Python just has a heavy "memoise this function please" lru cache as part of the stdlib.  Batteries included!