2023 Day 13 Part 2
I thought I was done, but I ended up having a flash of inspiration - you need to find a reflection where there is only one row that differs by only one char.  This we can calculate at the same time as we look at all the other rows to find the main reflection.  TBH if we only wanted to solve for part 2 we might even be able to stop ASAP but I decided I still wanted to know the main reflection.

Didn't actually take that long to spin up so I guess I could have got this star earlier had I been in the mood.