2023 Day 14
Part 1 was fairly trivial - just implement the tilt and load functions and away you go.  Used _far_ too much array chaining methods so I didn't have to really think about shuffling things by index, and used the transpose trick again to turn a vertical into a horizontal.  Happily '.', 'O', and '#' sort properly if you break the string on every '#' (e.g. '..O.#.O###O.' breaks into '..O.#', '.O#', '#', '#', 'O.') and this means tilting is "just" splitting and then sorting so that 'O' bubbles to the top.

Part 2 had me worried.  I implemented the other tilt directions easily enough (bit of head scratching to make sure I had the right combo of transpose and reverses for each direction, but I got there).  Obvs, ran it and it was too slow.  Went to reddit and saw someone mention "loops" and that gave me an idea to implement `spin` to find the loop and then work out how many more spins were needed.

Took me a while to get that right as you don't loop back to the first position, but to some intermediate position, and it took me time to realise that.  Got that working and it was fast enough, 'cos we only do 10 iterations to get a loop on the example (which repeats us back to the 3rd iteration giving loop period of 7) and my input only needed 168 spins to loop.  Reddit suggested that some people had loops of like 9 or something.  Still "loops" was the hint I needed.

Again, exploring the data is really the tip for the solution, not just hard thinking about the problem.