2023 Day 16 Part 1
Chipped away at this in bits and pieces over the day.  Lots of silly mistakes with confusing direction and applying that to the mirror (e.g. heading north means you've entered from the _south_).

Based on experience of the past days I'm already looking for loops and shortcircuting, so I hope that comes to help in part 2.