2023 Day 16 part 2
Ok, so I thought I'd need to optimise this more.  I had an idea that if I kept a map of all the things I'd seen across all the attempts then I could short circuit a given attempt if I came to that place.  However, that only works if you only have one beam, but we have loads of beams so what does it mean to say "I got to this point after energizing 10 spots in a previous attempt" if you've got 6 other beams on the go?  Can you do anything meaningful to calculate the result of the current attempt?  I don't think you can.  I originally thought I could calculate the diff between the energy at that point in the previous attempt and the energy at this point in the current attempt and then applying that diff to the final score of the previous amount.  Didn't work, obvs, so in the end, just ran the whole thing and it was fast enough.

Nothing on reddit suggested that there was a trick here (lots of neat visualisations though!).  I almost went straight to the Parallel gem as I assumed I'd get some benefit, but turns out, didn't need it.  Even with all my debug output it only took ~20s to get the solution and this thing is memory + object heavy.