2023 Day 17
Part 1 took longer than I'd have liked because of lots of silly off by one errors.  Crucially, I didn't really write any tests for this one until I was deep in implementation and that caused me a lot of pain.  Along with standard x + y confusion, and n,e,s,w confusion.  I initially started using the `SortedSet` that's now a gem but I think it expects objects to have unique ordering and if an object has the same ordering as something already in the set, it won't add it.  It took me a while to understand this and why I was getting weird results.  Switched to an array that I sort every time.

That's probably the cause of the slowness in my implementation - it took ~17mins to solve Part 1.

Part 2 didn't take too long to augment to understand the new movement constraints, but it took 2hrs to run, so I had to leave it overnight.  Definitely should have tried to optimise insertion to avoid sorting.