2023 Day 18 Part 2
Spelunked reddit for a bit to get some hints and they all spoke about shoelace formula, pick's theorem or green's theorem.  Shoelace seemed the easiest to understand so I went with that.  However,  I just wasn't getting the correct results no matter how much I poked and tweaked it.

Turns out, you can't just use shoelace theorem because our "perimeter" isn't a 0-width line, it's part of the area.  So we need to add that to the shoelace formula, and chuck in a +1 for good measure.

At least, that gets me the correct result for the example and input, but not for a small simple "rectangle with a chunk out of it" that I was trying my formula on.  I'm not _entirely_ sure what's going on here.  Is it a function of my rectangle not being irregular enough?  My formula says 41, but it should be 38.

I'm not really in the frame of mind to think this through, so I've ignored it.