2023 Day 19 Part 2
Really struggled with this as I was thinking about how to transform my first solution into something that dealt with ranges.  Realised that was a mistake and I basically needed to create a sankey diagram.  I neeeded a graph with:

    {x: 1..4,000, m: 1..4,000, a: 1..4,000, s: 1..4,000}

at the top level and each workflow is the next level, generating a new set of ranges per rule in the workflow.  e.g.

the first workflow in the example:

    in{s<1351:px,qqz}

creates a second level range split of:

    {x: 1..4,000, m: 1..4,000, a: 1..4,000, s: 1..1350}
    {x: 1..4,000, m: 1..4,000, a: 1..4,000, s: 1351..4_000}

where the first split points to whatever the workflow `px` creates, and the 2nd split points to whatever the workflow `qqz` creates.  We keep doing this until we've used up all the workflows.  Then we walk the graph to collect all the ranges that point to an 'A'. Then we multiply the ranges in each of those, and add them up.

In previous days we learned that part 2 is about optimisation, or that we need to explore the data to solve the problem in the data, not the general case.  In this day I learned that my "make a domain model that represents the problem and solve that" can sometimes get in the way of completely re-thinking what needs to be done.