2023 Day 21 Part 1
More pathfinding and grid walking.  Pretty straightforward with a set to keep things unique and nesw movement.  I used my visualiser and could see that going to 64 steps involved a lot of toggling of states.  I wonder if that'll be a concern in part 2?

Skipped day 20 because although it looked interesting, Day 21 seemed simpler.