2023 Day 21 Part 2
This was a nightmare.  I had no ideas at all, so scoured reddit for a few ideas, saw some really interesting visualisations that made me realise that the result could be calculated, but also that there was a quirk in the input that made the results easier to calculate vs. a version that solves the general case.

Scratched my head a bunch at this (see missteps in spec-2.rb and solution-2-wip.rb).  Ultimately after trying a bunch of ideas I came back to reddit and saw people talking about quadratic equations.  If you knew the count of possible plots at steps that hitting the "repeat" boundaries then you can solve a quadratic equation using [Lagrange Interpolation](https://www.geeksforgeeks.org/quadratic-interpolation/).  So I changed my solution-1 to expand the world and brute forced the data for 65, 196, 327 steps (the boundaries of 1 plot, 2 plots (9 tiles), 3 plots (25 tiles)).  This took ~2s, ~10mins, ~1hr:45mins but gave me the figures I needed and then I just coded up a solver.  It gave the right results for those values, and while it would have been nice to have 4 plots (49 tiles) to hand to compare, I don't think my ruby solution would be fast enough (I assume we're talking day-scale).

As the 3 plots version was running I whipped up a JS transliteration of my ruby brute-force one on the assumption it'd be faster.  It was _significantly_ slower because the Set implementation doesn't seem to work, so I was breaching the maximum size of it.

We'll see what Day 22 brings, and maybe I'll go back to Day 20, but I think I might be done.  At least on the Part 2s.