2023 Day 23
Part 1 was pretty fun - the slopes meant that a simplistic walk each step and collect paths to continue exploring at each junction would complete in reasonable time.  I got this one done fairly quickly.

Part 2 was trickier - slopes not meaning anything special make the search space much bigger.  I made the smallest change possible which was to ignore the special casing of the slopes and attempted to run the solution.  Obviously I don't know how long it would take, but it didn't look like it was going to complete any time soon.

My assumption was that I needed to prune the search space.  Watching the output of `debug_render` while I ran it made it clear to me it was very often searching the same parts of the graph over and over, just making different choices to explore the same space.  This made me attempt some caching strategies like not exploring a path if we'd been at the same spot before and it was longer.  That didn't work even on the example because an existing longer path to a point could be a dead-end that closed off other options for exploring and getting to the goal.  So I cached position and the choices available.  That worked on the example, and was fast on the input, but it wasn't exploring the full space (e.g. I got an answer that was too short).

I was stumped at this point so I explored Reddit a bit to see how other people were solving it.  Most people just had super efficient solutions that brute forced it.  I was sure there were some search space pruning options available though so I tried again.

I did think that efficiency had some efficiency options though.  First, I turned off `debug_render` - that definitely made it faster.  Then I extracted all the junctions and instead of stepping along the path one grid at at time, we just skip to the junction point and explore the choices available from it.  This is more efficient, but not enough to get a solution -- I had something running for ~5hrs using this and while it was finding longer paths than the caching attempt, it wasn't finding the longest option in reasonable time.

My cache was blocking exploration because different paths to a junction can mean you get the same choices available, but later choices will be different.  So a shorter path might still be worth exploring if later options would be different.  I _think_ there might be something worth exploring in fixing my cache to look at not just the choices available now, but the junctions not visited yet.

But, being lazy and confused at this point I went back to reddit and saw some talk about the shape of the graph if you looked at the connections between the junctions, so I made my solution generate graphviz of the junctions.  It is basically a square grid, and not very big, but that doesn't really help me -- the discussion of using this to prune options because the "perimeter" edges are directed didn't seem that clear to me (not least because I couldn't work out how to decide what "direction" the directed perimeter edges should be in - at least not from first principles while extracting the nodes, it's obvious when looking at the graph).

The only thing I did take from it is that the final junction that lets you get to the goal can be pruned to be a single choice -- the one that takes you to the goal.  Any other choice is going to block you from getting to the goal.  (This only holds true because there's only one way to the goal, if there were multiple final junctions it wouldn't apply).  Another "look at the data and build a solution for it, not the general case".

In the end, all the reddit posts spoke about just having an efficient solution.  I looked at a couple of solutions and after transliterating them in my head or de-golfing them it was clear they weren't doing any fancy pruning or anything.  They were fairly simple depth-first searches of the graph.  Just more efficient.

So I went back to my solution and spotted two major efficiency holes:

1. keeping all the paths around and iterating over them to find the longest one that's not done yet each step.  We only need to consider a few paths:
   1. the current longest one
   2. the one we're looking at right now
   3. the ones we'll create by hitting the junction at the end of the current path
   Obvs, 3 is a persisted set of all those generated at each step we haven't yet looked at.
2. stripping back the visited set from the paths - I was using this to make it easy to render the current path on the grid - but we don't need each path to have each grid reference, just the junction points.

Doing these two things meant I could run my brute force solution and explore the whole space in reasonable time.  It took about 8 minutes to find a longest path and happily, a longest path that got me a gold star.

Oof.  Didn't help that I was battling a cold and chipping away at this in 30 minute chunks over several days.

Still, here we are!  Day 23 complete!

I think I might attempt to at least get the silver star for the rest of the days.