2023 Day 22 Part 1
Actually quite fun and, being a part 1, fairly straight forward.  In 2021 there was a puzzle about 3d co-ordinates with beacons that absolutely did my head in so I was pleasantly surprised that this one wasn't that bad at all.

As ever, I just modelled the problem and build objects that made the process simple as I could focus on one small part of the problem at a time.  Sort of fell apart on testing things though.

It's wildly inefficient, but no point trying to optimise until part 2 because who knows how it'll change the puzzle.